backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-multipart
python-jose
pdfplumber
google-genai
pydantic-settings
python-dotenv
email-validator
bcrypt